Fixed the shop special item not spawning
So the problem was that first off, the spawn spot was on the floor,
which was making the item not spawn because it was blocked by the
credits shelf.

Secondly, if the item was spawned, it wouldn't ever be removed because
it wasn't being assigned a TID, so the script mistakenly removed the map
spot instead, causing nothing to spawn there ever again.