Lots of Shop Special fixes
Fix all of the references to PlayerNumber in outpost menus where a
boolean value is expected. This could cause players to have their frozen
flag arbitrarily removed or set in a net game where a player other than
player 1 enters or leaves a menu.

Remember to unfreeze players in the Shop Special menu if someone else
buys the item. In the case that the item was already bought, tell the
player when the next restock will occur.