A more situational autosave system
The new system will try to avoid autosaving if you've taken damage in
the last 20 seconds or have very low health, to prevent getting stuck in
death loops or autosaving in the middle of firefights. If such is the
case, it will wait until you are at better health and are out of the
fight before saving.

It will also take opening the menu without menu-freeze enabled to be
outside of a firefight and will try to save then as well, if your health
isn't critical.