Extras: Big partial update
The most important aspect of this update is that the gibbers were
compressed into a more readable and versatile form. As a result, blood
colors are now appropriately handled by gibbers without special
requirements, and bloody chunks will now spray across the room,
preserving the original actor's velocity, when they're hit with a
powerful weapon and gibbed and sent flying.