Tweak monster vit. formula, fix non-RLA mission rewards
The monster vitality formula needs to give exactly 10x
health at 1,000 vitality stat, so calculate it to 9x and
add the result to the base health so we wind up with 10x
instead.

If we don't have RLA loaded, we need to swap boots for
armor, and modpacks for stims so we don't pick from
empty categories.