Reinforcements: Spawn them more consistently
This code will make them spawn on the floor, so that
they don't spawn in mid-air or fail to spawn in places
that are otherwise open for spawning due to minor floor
height variations.

We also try again up to 3 times if we fail, tightening
the circle in which they will spawn each time, so that
reinforcements won't be incredibly scarce in close
quarters, since melee enemies are less threatening from
afar.