Completely change the reward code
We don't bias the items anymore. What we do now instead
is create a secondary list of items, sort by their
price, then choose the item based on the sorted list,
least to most expensive. This completely ignores
category weights in preference for the item's actual
credit value. As a result, items that should be easier
to get now are.