Fixes with the damagetype system, as it was terribly broken
The RLA player classes now account for the subtypes of
damage that can otherwise be caused to the player. We
also cannot clear the damagetype that was last caused
because some asynchro-stupidity of some running scripts
means that we never actually get to *see* what type of
damage was caused to us otherwise, as it gets cleared
before the HUD ever gets to touch it.