New mechanics for the Environmental Hazard event
Upon entering the map, it will have a hazard level of 1 through 5. Based on
that level, there will also be a toxicity percent. A radiation neutralizer
device will spawn.

To reduce the toxicity level of the area, the neutralizer must be supplied
liquid fuel, which will also spawn randomly around the map. You must keep
supplying fuel to the generator at regular intervals until it reduces the
hazard level and toxicity. Once the toxicity reaches 0%, the radiation is
cleared, and the event will end.