Completely overhaul the way maps are stored and handled
The levels array is now not only populated on the fly, but numbered on the
fly as well, so map packs that don't define level numbers for their maps
will still function. This also means that secret maps will be numbered and
have monster levels close to the maps from which they exited.

The Outpost is now also stored in the levels array and, while currently
hardcoded by name, determined at run-time by the presence of an actor in
the level. This will allow not only events to be decided on the Outpost in
the same run as the other maps, but also allow possibly multiple Outpost
and Arena maps and further mapping expansions later on.