Ensure that GetMonsterID is valid in MonsterInitStats
An attempt to fix the runaway bug in MonsterInitStats, which could
/possibly/ be caused by an invalid monster ID access when re-stat'ing
the monster, since static access is handled through stdlib functions that
could be causing the script to explode upon trying to find an invalid
static address from reading garbage.

This is definitely sign of a much deeper-rooted problem that's going to
require further examination, but here's a start at least.