Add an option to make auto-sell fall back to auto-deposit.
When you're first starting out at rank 0, you can't yet sell items outside of the Outpost, but that doesn't mean you have to carry them around until you rank up or return to the Outpost; you can still deposit them in the locker.

This commit adds an option that, if turned on, automates this. When picking up an item that you've set to auto-sell, but can't actually sell it at the moment, you'll automatically deposit it in the locker instead. Normal auto-sell behavior still applies if you're at least rank 1 or are in the Outpost.