Fix destroyed turret parts spawning in unreachable places.
Previously, when the player looted a destroyed turret, the resulting turret parts would appear in a circle around it. If it was next to a wall, some of them would appear *inside* the wall! Similarly, some of the parts might have fallen off a ledge, or climbed up a ledge, to a spot that is visible but unreachable.

This commit has the game's physics engine make sure that the turret parts don't wind up anywhere you can't easily reach.