Add a sanity check to fix issue #68.
`kill monsters` was causing the turret to become “destroyed” without altering its health, making it impossible to deploy or repair. This works around that problem by checking every tic if its health is full, and if so, clearing the destroyed flag.

This was also happening very rarely under some other circumstances, but I'm not sure what. This will presumably fix those as well.