Stability changes, DRLA checks
Many more DRLA checks to prevent calls to nonexistent objects. Somewhere
in all of this, a memory corruption bug was fixed. A full day's testing
performed, using GZDoom builds 1651 and 1664, with no script
terminations.

Significantly reduced the volume of the Outpost song to bring it in line
with average MIDI volume.