Fixed regression in Mission BBS
Mission difficulties are supposed to be restricted by rank, this was
missed when switching BBS input to a consistent input method.
Hide HUD elements when a menu is open. This already happened to some
degree with the credits/modules display, but was broken on the DRLA HUD,
and is now corrected.