Fixed Assassination missions when playing with  LegenDoom
Assassination missions were checking for the wrong monster classnames
when legendoom was loaded.
DoomRPG is supposed to spawn some extra enemies on map start based on
difficulty - this script was always being killed for too many
iterations.  On correcting that I found the code was having a very hard
time finding suitable locations - this was due to it being very
conservative in case it tries to spawn a spider mastermind. Removed the
mastermind from the possible spawned monsters and lowered the size
requirements.
Fixed Credits not showing when in the shop.
More GUI work, still got a long way to go!