LegenDoom HUD added
Added a LegenDoom HUD similar to the one in the mod proper.
Displays the weapons you have and what status effect that weapon has, if any.
Added the usual cvars for controlling the position of the HUD element, and added the sliders to the HUD menu.
Also has support for HUD preview.
Corrected alignment of the Fists icon in the shop for LegenDoom.

Added monster drops to the list of items a magnetic stim will attract.
This was an attempt to fix a bug with the magnetise skill with LegenDoom, but the bug was caused by LegenDoom itself.
This bug has been fixed and will be rolled out with the next LegenDoom release.