Stupid hack
For some weird reason, in the latest versions of ZDoom, garbage is being appended to the end of this list and I can't figure out the cause for it. Until I do, which I'm not really holding my breath on, I'll just use a stupid hack to get around it.